refactor: 오타 수정 및 log 삭제
