+++ src/main/java/back/vybz/paymentservice/subscription/vo/response/ResponseBuskerProfileVo.java
@@ -1,4 +1,4 @@
-package back.vybz.paymentservice.subscription.vo.respomse;
+package back.vybz.paymentservice.subscription.vo.response;
 
 import lombok.Builder;
 import lombok.Getter;
